Se añadio la navegación entre pantallas y se añadio la pantalla de registrar usuario
