+++ package.json
@@ -9,11 +9,16 @@
     "web": "expo start --web"
   },
   "dependencies": {
+    "@react-navigation/native": "^6.1.3",
+    "@react-navigation/native-stack": "^6.9.9",
+    "@react-navigation/stack": "^6.3.12",
     "expo": "~47.0.12",
     "expo-status-bar": "~1.4.2",
     "react": "18.1.0",
     "react-dom": "18.1.0",
     "react-native": "0.70.5",
+    "react-native-safe-area-context": "4.4.1",
+    "react-native-screens": "~3.18.0",
     "react-native-web": "~0.18.9"
   },
   "devDependencies": {
